+++ NEWS.md
@@ -7,7 +7,7 @@ Cite all releases with doi [10.5281/zenodo.5118872](https://doi.org/10.5281/zeno
 which always resolves to the latest release.
 
 
-# matsbyname 0.6.5 (2023-09-16)
+# matsbyname 0.6.5 (2023-09-16) [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.8353760.svg)](https://doi.org/10.5281/zenodo.8353760)
 
 * Remove deprecated `column` argument on `vec_from_store_byname()`.
 * Remove deprecated argument `matrix.class` in many places.